feat(categoryHandler): permite uso da classe por todas as Categorias         adiciona parâmetro modifiedSince à requisição visando execução por job         adiciona loop para get na API ao método da classe
